refactor: simplify IO
Renamed IReadOnlyAccess to SeekableReadOnlyByteChannel.
Renamed ReadOnlyAccessFile to SeekableReadOnlyFile.
Renamed InputStreamReadOnlyAccessFile to SeekableReadOnlyInputStream.
The naming refers to the ones in the nio package.
Channel refers to the underlying data provider.
Seekable refers to the fact that those providers will allow to set position arbitrarily.

Changed ReadOnlyAccessFile to not extend RandomAccessFile and use it in a field instead.
This brings a similar behavior as SeekableReadOnlyInputStream.

Removed ReadOnlyAccessInputStream which was a thin wrapper around a IReadOnlyAccess.
ComprDataIO now uses the underlying archive's channel directly.

Removed the ability to create a RandomAccessStream from a RandomAccessFile.

Renamed the impl package to volume.

BREAKING CHANGE: name of classes have changed